realEndPage구할 때 total과 perPagePageNum 모두 int형이라서 나누기를 하면 정수가 나오는 바람에 (정수/정수=정수) total=512개 perPagePageNum=10개 라면 52페이지까지 나와야하는데 51페이지까지만 나오는 현상이 발견되었다. 그래서 perPagePageNum을 강제형변환(double)하여 이 문제를 해결!

+++ src/main/java/org/bamboo/ilovehub/domain/PageCreator.java
@@ -26,7 +26,8 @@ public class PageCreator {
 		this.endPage=(int)Math.ceil(cri.getPage()/(double)perPagePageNum)*perPagePageNum;
 		this.startPage=endPage-(perPagePageNum-1);
 		
-		this.realEndPage = (int)(Math.ceil(total/perPagePageNum));
+		this.realEndPage = (int)(Math.ceil(total/(double)perPagePageNum));
+		//System.out.println("이해불가:"+Math.ceil(total/(double)perPagePageNum)+", total:"+total+", perPagePageNum:"+perPagePageNum);
 		if(realEndPage<endPage) {
 			this.endPage=realEndPage;
 		}
